chore: refactor structure for auth
