+++ src/app.ts
@@ -1,10 +1,10 @@
-import express, { type Express } from 'express';
+import express from 'express';
 import swaggerUi from 'swagger-ui-express';
 import cors from 'cors';
 import cookieParser from 'cookie-parser';
 import expressWinston from 'express-winston';
 import helmet from 'helmet';
-import routes from './routes';
+import routes from './http/routes';
 import swaggerDocument from './docs';
 import {
   requestHandler,
@@ -12,7 +12,7 @@ import {
   requestLogger,
 } from './http/middlewares';
 
-const app: Express = express();
+const app = express();
 
 app.use(helmet());
