+++ src/http/middlewares/index.ts
@@ -1,5 +1,5 @@
 import { requestHandler } from './requestHandler';
 import { errorHandler } from './errorHandler';
-import { requestLogger } from '../../lib/winston';
+import { requestLogger } from '../../lib/winston/logger';
 
 export { requestHandler, errorHandler, requestLogger };
