+++ src/lib/prisma/db.ts
@@ -1,17 +1,14 @@
 import { PrismaClient } from '@prisma/client';
 
-const prisma = new PrismaClient({
+export const db = new PrismaClient({
   log: process.env.NODE_ENV !== 'production' ? ['warn', 'error'] : [],
   errorFormat: process.env.NODE_ENV !== 'production' ? 'pretty' : 'colorless',
 });
 
-prisma
-  .$connect()
+db.$connect()
   .then(() => {
     console.log('📦 Successfully connected with database');
   })
   .catch((error) => {
     console.log('❌ Error connecting to database', error);
   });
-
-export default prisma;
