+++ src/server.ts
@@ -2,7 +2,7 @@ import 'dotenv/config';
 
 import './env';
 import app from './app';
-import 'src/lib/prisma';
+import 'src/lib/prisma/db';
 
 app.listen(process.env.SERVER_PORT || 3001, () => {
   console.log(
